app/requirements.txt
fastapi
python-multipart
jinja2
uvicorn
python-jose
python-jose[cryptography]
